plate2people/backend/requirements.txt
Django>=6.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
django-channels>=4.1.0
channels-redis>=4.1.0
python-socketio>=5.10.0
python-socketio[aiohttp]>=5.10.0
django-anymail>=10.0
razorpay>=1.4.1
cloudinary>=1.36.0
python-magic>=0.4.27
drf-spectacular>=0.27.0
celery>=5.3.0
gunicorn>=21.2.0
python-decouple>=3.8
redis>=5.0.0
